refactor: GolfPose2DPanel 컴포넌트 분리 - 기존 : 골프자세 2D & 8개의 스윙 이미지 렌더링 - 변경 : GolfPose2DPanel → 골프자세 2D         GolfPoseSwingImageGrid → 8개의 스윙 이미지 그리드
